NEW!
Updated the Command list to include a CREATETABLE command.

Updated the Database.java file to include a method for dropping a table
if it exists.

Added a createRow, createTable, and queryTable method that conforms to
Query.  Also added a dropTable that conforms to the Database.java
abstract.

Starting to update query to funnel all getQuery requests through a
single method that will return the needed string as per the command.

Updated the QueryBuilder.java to subclass Query to get rid of stupid
duplicates.  Now to use, you do the same as before without needing to
buildQuery(), and getQuery()